Finite State Machine architecture now fully in place.
In this commit the FSMState was pulled out of the GameState and placed
into the BackendState. It should have been like this before.
The backend must be aware of the FSMState of the game in order
to provide the right events.

Another change that was made was that now we only process one
game event per frame. However, we do not consider an SDL
event not understood by the game to be an event. We keep
polling until either we get an SDL event that we can convert into
a game event or S.NoEvent.