Refactored the GLM monad.
The GLM monad is now not just a wrapper around IO.
It is a function from GLSLState -> IO a. This means that
we can get all the GLSL state upfront and inject into
all GLM calculations.

This required changing the type of runGameM a bit but not
so that it was completely disgusting. We still have
a really nice encapsultion going where GameM calculations
can know nothing about annoying backend state except inside
values of the GLM monad.