drawText now only takes the width argument, not height.
This required quite a lot of work. The main problem was that
the Cairo function 'textExtents' was the only way of calculating what
the height of a piece of rendered text would be. It was "trapped" inside
the Render monad. I ended up writing a function called "runWithoutRender"
that would render a Cairo picture to a 1x1 image buffer and then throw it away, but
return the result of the Render monad inside the IO monad.

I then had to generalise the types of a whole bunch of functions in
moudles GraphicsGL and Graphics in order to accomplish what I wanted
from return type IO () or GLM () to IO a or GLM a respectively.